O7E-70 Refactor Header and Navigation components to improve routing and introduce navigationLinks structure
