+++ src/components/Navigation.vue
@@ -24,110 +24,21 @@ import { computed } from 'vue';
 import { useRouter } from 'vue-router';
 import { useDrawer } from '@/stores/drawer'
 import { UserRole } from '@/enums/enums';
-
-
-const drawer = useDrawer();
-
-interface NavigationGroup {
-  Name: string;
-  Links: NavigationLink[]
-}
-
-interface NavigationLink {
-  Name: string;
-  AllowRoles: UserRole[];
-  Route: string;
-}
-
-const items: NavigationGroup[] = ([
-  {
-    Name: "Vozilo",
-    Links: [
-
-      {
-        Name: "Promjena vlasništva",
-        AllowRoles: [UserRole.HAK],
-        Route: '/ownership-change'
-      },
-      {
-        Name: "Novo vozilo",
-        AllowRoles: [UserRole.HAK],
-        Route: '/new-vehicle'
-      },
-      {
-        Name: "Tehnički pregled",
-        AllowRoles: [UserRole.HAK],
-        Route: '/technical-check'
-      },
-      {
-        Name: "Odjava vozila",
-        AllowRoles: [UserRole.HAK],
-        Route: '/vehicle-deregistration'
-      },]
-  },
-  {
-    Name: "Službenici",
-    Links: [
-      {
-        Name: "Pregled službenika",
-        AllowRoles: [UserRole.Admin],
-        Route: '/officer-overview'
-      },
-      {
-        Name: "Novi službenik",
-        AllowRoles: [UserRole.Admin],
-        Route: '/new-officer'
-      },
-    ]
-  },
-  {
-    Name: "Promet",
-    Links: [
-      {
-        Name: "Prometna dozvola",
-        AllowRoles: [UserRole.Osoba],
-        Route: '/traffic-license'
-      },
-      {
-        Name: "Vozačka dozvola",
-        AllowRoles: [UserRole.Osoba],
-        Route: '/driver-license' 
-      },
-      {
-        Name: "Uređaji",
-        AllowRoles: [UserRole.Osoba],
-        Route: '/devices'
-      },
-      {
-        Name: "Vozila",
-        AllowRoles: [UserRole.Osoba, UserRole.Firma],
-        Route: '/vehicles'
-      }
-    ]
-  },
-  {
-    Name: "Akcije",
-    Links: [
-    {
-        Name: "Novi korisnik",
-        AllowRoles: [UserRole.SuperAdmin],
-        Route: '/new-user'
-      }
-    ]
-  },
-
-
-]);
+import { navigationLinks, type NavigationGroup } from '@/layouts/links';
 
 const currentUserRole = UserRole.Osoba;
+
+const drawer = useDrawer();
 const router = useRouter();
-const filteredItems = computed(() => {
-  items.forEach(item => {
+
+const filteredItems = computed<NavigationGroup[]>(() => {
+  navigationLinks.forEach(item => {
     item.Links = item.Links.filter(link => link.AllowRoles.includes(currentUserRole))
   })
 
-  return items.filter(item => item.Links.length != 0)
+  return navigationLinks.filter(item => item.Links.length != 0)
 })
+
 
 </script>
